requirements.txt
EasyProcess==0.1.6
chardet==2.3.0
praw==2.1.20
python-dateutil==2.2
python-firebase==1.2
requests==2.3.0
six==1.8.0
update-checker==0.11
virtualenv==12.0.7
